pyproject.toml
[tool.poetry]
name = "project-garmin-backend"
version = "0.1.0"
description = "Backend for Local Personal AI Assistant (Garmin) using FastAPI"
authors = ["Igor Podlecki <igor@example.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlmodel = "^0.0.22"
sqlalchemy = "^2.0.30"
chromadb = "^0.5.5"
pypdf = "^4.2.0"
fpdf2 = "^2.7.9"
python-multipart = "^0.0.9"
sentence-transformers = "^3.0.1"
pillow = "^10.4.0"
pytesseract = "^0.3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

